+++ include/GafferUI/PathColumn.h
@@ -89,8 +89,11 @@ class GAFFERUI_API PathColumn : public IECore::RefCounted, public Gaffer::Signal
 				const IECore::ConstDataPtr &value = nullptr,
 				const IECore::ConstDataPtr &icon = nullptr,
 				const IECore::ConstDataPtr &background = nullptr,
-				const IECore::ConstDataPtr &toolTip = nullptr
-			)	:	value( value ), icon( icon ), background( background ), toolTip( toolTip ) {}
+				const IECore::ConstDataPtr &toolTip = nullptr,
+				const IECore::ConstDataPtr &sortValue = nullptr
+			)	:	value( value ), icon( icon ), background( background ), toolTip( toolTip ),
+					sortValue( sortValue ) {}
+
 			CellData( const CellData &other ) = default;
 
 			/// The primary value to be displayed in a cell or header.
@@ -119,6 +122,9 @@ class GAFFERUI_API PathColumn : public IECore::RefCounted, public Gaffer::Signal
 			///
 			/// - StringData
 			IECore::ConstDataPtr toolTip;
+			/// Used to determine sort order. If not specified, `value` is
+			/// used for sorting instead.
+			IECore::ConstDataPtr sortValue;
 
 			private :